Fix clippy lints

+++ src/lib.rs
@@ -63,7 +63,7 @@ pub use transport::{unencrypted, Duplex, RoutedFrame};
 /// Maximum message (packet payload) length for Brontide protocol
 pub const BRONTIDE_MSG_MAX_LEN: usize = u16::MAX as usize;
 /// Maximum message (packet payload) length for Brontozaur protocol
-pub const BRONTOZAUR_MSG_MAX_LEN: usize = 0xFFFFFF as usize;
+pub const BRONTOZAUR_MSG_MAX_LEN: usize = 0xFFFFFF;
 
 /// Trait used by different address types (transport-, session- and
 /// presentation-based) for getting scheme part of the URL
